Optimized Atom Deactivation.
Removing atoms from the global stack is now scheduled.
This means that they will all be removed at the same time which is
more efficient and will not require searching for the index of the atom
to remove.

This has cut down the benchmarks for clearing all rows by almost half.